pyproject.toml
[project]
name = "flow-matching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lightning>=2.5.0.post0",
    "matplotlib>=3.10.0",
    "numpy",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.0",
    "torchdiffeq",
    "tqdm",
    "wandb>=0.19.1",
    "zuko",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'"},
]
torchvision = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'"},
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
